test: share a user_contact_info scenario across data-enum tests
The `User { id, name, contact: ContactInfo }` model with a data-carrying
`ContactInfo { Email, Phone }` enum was redefined inline in seven tests
across four files. Extract it into a `user_contact_info` scenario and have
those tests consume it, so they now also run in both id modes.
