+++ crates/toasty-driver-integration-suite/src/scenarios.rs
@@ -20,6 +20,7 @@ pub mod two_models;
 pub mod two_models_unique_title;
 pub mod user_account_subscription;
 pub mod user_comment_article;
+pub mod user_contact_info;
 pub mod user_name_email;
 pub mod user_org_project_todo;
 pub mod user_post_comment;
