+++ crates/toasty-driver-integration-suite/src/tests/embed_enum_filter.rs
@@ -2,27 +2,9 @@ use crate::prelude::*;
 
 /// Filtering by a data-carrying enum value using a SQL WHERE clause.
 /// DynamoDB does not support arbitrary filter predicates, so this is SQL-only.
-#[driver_test(requires(scan))]
+#[driver_test(id(ID), requires(scan), scenario(crate::scenarios::user_contact_info))]
 pub async fn filter_data_enum(t: &mut Test) -> Result<()> {
-    #[derive(Debug, PartialEq, toasty::Embed)]
-    enum ContactInfo {
-        #[column(variant = 1)]
-        Email { address: String },
-        #[column(variant = 2)]
-        Phone { number: String },
-    }
-
-    #[derive(Debug, toasty::Model)]
-    #[allow(dead_code)]
-    struct User {
-        #[key]
-        #[auto]
-        id: uuid::Uuid,
-        name: String,
-        contact: ContactInfo,
-    }
-
-    let mut db = t.setup_db(models!(User)).await;
+    let mut db = setup(t).await;
 
     User::create()
         .name("Alice")
@@ -53,27 +35,9 @@ pub async fn filter_data_enum(t: &mut Test) -> Result<()> {
 }
 
 /// Filtering by variant alone (discriminant-only check) using `is_{variant}()`.
-#[driver_test(requires(scan))]
+#[driver_test(id(ID), requires(scan), scenario(crate::scenarios::user_contact_info))]
 pub async fn filter_data_enum_by_variant(t: &mut Test) -> Result<()> {
-    #[derive(Debug, PartialEq, toasty::Embed)]
-    enum ContactInfo {
-        #[column(variant = 1)]
-        Email { address: String },
-        #[column(variant = 2)]
-        Phone { number: String },
-    }
-
-    #[derive(Debug, toasty::Model)]
-    #[allow(dead_code)]
-    struct User {
-        #[key]
-        #[auto]
-        id: uuid::Uuid,
-        name: String,
-        contact: ContactInfo,
-    }
-
-    let mut db = t.setup_db(models!(User)).await;
+    let mut db = setup(t).await;
 
     User::create()
         .name("Alice")
@@ -270,27 +234,9 @@ pub async fn filter_enum_variant_with_partition_key(t: &mut Test) -> Result<()> 
 /// Creates records with different data-carrying enum variants and retrieves them
 /// by primary key, verifying enum values round-trip correctly. This exercises
 /// the same create + read path on all drivers including DynamoDB.
-#[driver_test]
+#[driver_test(id(ID), scenario(crate::scenarios::user_contact_info))]
 pub async fn create_and_get_data_enum(t: &mut Test) -> Result<()> {
-    #[derive(Debug, PartialEq, toasty::Embed)]
-    enum ContactInfo {
-        #[column(variant = 1)]
-        Email { address: String },
-        #[column(variant = 2)]
-        Phone { number: String },
-    }
-
-    #[derive(Debug, toasty::Model)]
-    #[allow(dead_code)]
-    struct User {
-        #[key]
-        #[auto]
-        id: uuid::Uuid,
-        name: String,
-        contact: ContactInfo,
-    }
-
-    let mut db = t.setup_db(models!(User)).await;
+    let mut db = setup(t).await;
 
     let alice = User::create()
         .name("Alice")
